+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## HEAD
+
+- add website utility class
 
 - Added changelog